requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
attrs==23.1.0
av==15.1.0
beautifulsoup4==4.14.2
bitsandbytes==0.48.1
blessed==1.22.0
blobfile==3.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
DALL-E==0.1
datasets==4.0.0
debugpy==1.8.16
decorator==5.2.1
decord==0.6.0
deepspeed==0.18.0
dill==0.3.8
distro==1.9.0
einops==0.8.1
evaluate==0.4.5
exceptiongroup==1.3.0
executing==2.2.1
filelock==3.19.1
flash_attn==2.8.3
fonttools==4.60.1
frozenlist==1.7.0
fsspec==2025.3.0
ftfy==6.3.1
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.45
gpustat==1.1.1
h11==0.16.0
hf-xet==1.1.9
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
iniconfig==2.1.0
iopath==0.1.10
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
lazy_loader==0.4
loguru==0.7.3
lxml==6.0.1
MarkupSafe==3.0.2
matplotlib==3.10.7
matplotlib-inline==0.1.7
mpi4py==4.1.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
multiprocess==0.70.16
mypy==1.17.1
mypy_extensions==1.1.0
nest_asyncio==1.6.0
networkx==3.5
ninja==1.13.0
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.580.65
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
openai==1.107.0
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.2
parso==0.8.5
pathspec==0.12.1
peft==0.17.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.3.0
pip==25.2
platformdirs==4.4.0
pluggy==1.6.0
portalocker==3.2.0
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==6.32.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==21.0.0
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.5
PySocks==1.7.1
pytest==8.4.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.2.0
qwen-vl-utils==0.0.11
regex==2025.9.1
requests==2.32.5
sacrebleu==2.5.1
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.1
sentry-sdk==2.37.0
setuptools==78.1.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
sqlitedict==2.1.0
stack_data==0.6.3
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.9.9
timm==1.0.19
tokenizers==0.22.0
torch==2.8.0
torchmultimodal-nightly==2024.4.1
torchvision==0.23.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
triton==3.4.0
trl==0.22.2
typing_extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
wandb==0.21.3
wcwidth==0.2.13
wheel==0.45.1
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
